+++ CHANGES.md
@@ -3,6 +3,12 @@
 This document describes the relevant changes between releases of the OCM API
 SDK.
 
+## 0.1.231 Dec 22 2021
+
+- Update to model 0.0.165:
+  - Move version gates to be top level resource.
+  - Add version raw id prefix to version gates.
+
 ## 0.1.230 Dec 21 2021
 
 - Update to model 0.0.163: